src/doc-qa/ingestion/requirements.txt
google-cloud-aiplatform>=1.38.0
google-cloud-bigquery>=3.13.0
google-cloud-storage>=2.14.0
google-cloud-secret-manager>=2.18.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
elasticsearch>=8.12.0
